+++ dnsimple/services_domains_test.go
@@ -47,7 +47,7 @@ func TestServicesService_AppliedServices(t *testing.T) {
 		t.Errorf("DomainServices.AppliedServices() expected to return %v services, got %v", want, got)
 	}
 
-	if want, got := 1, services[0].ID; want != got {
+	if want, got := int64(1), services[0].ID; want != got {
 		t.Fatalf("DomainServices.AppliedServices() returned ID expected to be `%v`, got `%v`", want, got)
 	}
 	if want, got := "wordpress", services[0].SID; want != got {
